requirements.txt
# Core requirement for both Google AI and Vertex AI
google-genai>=1.43.0

# Optional but recommended: For automatic video duration detection
yt-dlp

# Note: For Vertex AI, you also need:
# 1. Google Cloud SDK installed: https://cloud.google.com/sdk/docs/install
# 2. Authentication: gcloud auth application-default login
# 3. Project setup: gcloud config set project YOUR_PROJECT_ID
